mocha and istanbul are functional
