+++ test/test.js
@@ -1,5 +1,5 @@
 
-import constructorTest  from './unit/constructorTest';
-import flounderTest     from './unit/flounderTest';
-import utilsTest        from './unit/utilsTest';
-import versionTest      from './unit/versionTest';
+import flounderTest  from './unit/flounderTest';
+// import flounderTest     from './unit/flounderTest';
+// import utilsTest        from './unit/utilsTest';
+// import versionTest      from './unit/versionTest';
